Fix infinite loop in Thread.sleep

+++ Native/Src/flint_native_thread_class.cpp
@@ -38,8 +38,8 @@ static void nativeCurrentThread(FlintExecution &execution) {
 
 static void nativeSleep0(FlintExecution &execution) {
     uint64_t startTime = FlintAPI::System::getNanoTime() / 1000000;
-    uint64_t millis = execution.stackPopInt64();
-    while(((FlintAPI::System::getNanoTime() / 1000000) - startTime) < (millis - 100)) {
+    int64_t millis = execution.stackPopInt64();
+    while((int64_t)((FlintAPI::System::getNanoTime() / 1000000) - startTime) < (millis - 100)) {
         FlintAPI::Thread::sleep(100);
         if(execution.hasTerminateRequest())
             throw &execution.flint.newInterruptedException(*(FlintString *)0);
